+++ ChangeLog.md
@@ -3,6 +3,7 @@
 Real men use only pure calculus
 
 - v? (in progress)
+  - Add webassembly/emscripten support for running lci in a browser.
   - Add auto-completion of aliases and system commands.
 
 - v1.0 (15 Jun 2021)